requirements.txt
streamlit>=1.31.0
pdfplumber>=0.10.3
python-dotenv>=1.0.0

# LangChain + FAISS
langchain>=0.1.0,<0.3.0
langchain-community>=0.0.10,<0.3.0
faiss-cpu>=1.7.4

# Embeddings
sentence-transformers>=2.2.2,<5.0.0

# HF inference on CPU (pin Transformers to 4.x to avoid breaking sentence-transformers)
transformers>=4.41.0,<5.0.0
tokenizers<0.22
huggingface-hub<0.31
accelerate<1.0.0
sentencepiece>=0.2.0

# Torch CPU
torch>=2.1.0
